refactor: gather all predicate rewriting into a single file
